refactor user setting component and add more loaders
